接口提交

+++ src/main/java/com/tencent/wxcloudrun/dao/ActivityDetailMapper.java
@@ -27,7 +27,7 @@ public interface ActivityDetailMapper {
      * @author lf
      * @date 2023/08/09
      **/
-    int delete(int id);
+    int delete(Long id);
 
     /**
      * 更新
